backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
email-validator
python-dotenv>=1.0.0
pwdlib[argon2]
cryptography
PyJWT
python-multipart
rapidfuzz
pinecone
google-generativeai
google-genai
requests
langgraph>=0.2.0
langchain-openai
langchain-google-genai
motor
google-cloud-speech
elasticsearch
websockets
google-auth

# RAG Pipeline Dependencies
tiktoken>=0.5.0
aiofiles>=23.0.0
openai>=1.0.0

# Background Jobs
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0

# Optional Vector Stores
# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
greenlet>=3.0.0
asyncpg>=0.29.0
qdrant-client>=1.7.0

# Optional Document Loaders
boto3>=1.34.0
pypdf>=4.0.0

# Optional Local Embeddings
sentence-transformers>=2.2.0
cel-python>=0.1.5
beautifulsoup4>=4.12.0
